refactor bottom bar and cancel model
